+++ engine/interpreter/src/main/scala/pl/touk/nussknacker/engine/sql/TimestampUtils.scala
@@ -9,8 +9,4 @@ object TimestampUtils {
     Timestamp.from(localDateTime.atZone(ZoneId.systemDefault()).toInstant)
   }
 
-  def toLocalDateTime(t: Timestamp): LocalDateTime = {
-    LocalDateTime.ofInstant(t.toInstant, ZoneId.systemDefault())
-  }
-
 }